pyproject.toml
[project]
name = "fastapi-large-app-template"
version = "1.0.0"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "fastapi ~= 0.121.1",
    "uvicorn[standard] ~= 0.30.0",
    "gunicorn ~= 22.0.0",
    "SQLAlchemy ~= 2.0.44",
    "alembic ~= 1.17.1",
    "asyncpg ~= 0.29.0",
    "psycopg2-binary ~= 2.9.11",
    "redis ~= 6.0.0",
    "PyJWT ~= 2.9.0",
    "bcrypt >= 4.0.1, < 5.0.0",
    "passlib >= 1.7.4, < 2.0.0",
    "fastapi-limiter ~= 0.1.6",
    "pydantic ~= 2.8.0",
    "email-validator ~= 2.2.0",
    "python-decouple == 3.8",
    "openai ~= 1.12.0",
]

[project.optional-dependencies]
test = [
    "pytest ~= 8.0.0",
    "pytest-asyncio ~= 0.23.0",
    "httpx ~= 0.27.0",
]
dev = [
    "ruff ~= 0.8.0",
    "pre-commit ~= 4.0.0",
    "mypy ~= 1.18.2",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
explicit_package_bases = true
disallow_untyped_defs = false
check_untyped_defs = true
exclude = ["app/alembic/versions"]
pretty = true
show_error_codes = true
show_error_context = true

plugins = [
  "pydantic.mypy",
  "sqlalchemy.ext.mypy.plugin"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["app/alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]
